Refactor StoryMerger to multiple files
